# [PERF] stock_account, purchase_stock: Speedup pickings