--- a/addons/stock_account/models/stock_valuation_layer.py
+++ b/addons/stock_account/models/stock_valuation_layer.py
@@ -6,7 +6,7 @@ from odoo.exceptions import UserError
 from odoo.tools import float_compare, float_is_zero
 
 from itertools import chain
-from odoo.tools import groupby
+from odoo.tools import groupby, OrderedSet
 from collections import defaultdict
 
 
@@ -67,6 +67,8 @@ class StockValuationLayer(models.Model):
     def _validate_accounting_entries(self):
         am_vals = []
         aml_to_reconcile = defaultdict(set)
+        move_ids = OrderedSet()
+        svl_move_list = defaultdict(int) 
         for svl in self:
             if not svl.with_company(svl.company_id).product_id.valuation == 'real_time':
                 continue
@@ -75,7 +77,16 @@ class StockValuationLayer(models.Model):
             move = svl.stock_move_id
             if not move:
                 move = svl.stock_valuation_layer_id.stock_move_id
-            am_vals += move.with_company(svl.company_id)._account_entry_move(svl.quantity, svl.description, svl.id, svl.value)
+            move_ids.add(move.id)
+            svl_move_list[svl.id] = move.id
+        
+        moves = self.env['stock.move'].browse(move_ids)
+        move_directions = moves._get_move_directions()
+        for svl in self:
+            linked_move = moves.browse(svl_move_list[svl.id])
+            if linked_move:
+                am_vals += linked_move.with_context(move_directions=move_directions).with_company(svl.company_id)._account_entry_move(svl.quantity, svl.description, svl.id, svl.value)
+
         if am_vals:
             account_moves = self.env['account.move'].sudo().create(am_vals)
             account_moves._post()