### Description:

Validating a picking with serial numbers can be slow. This is caused by the method `_account_entry_move` which was made to be called one by one on each valuation layer. The issue is each time this method is called, it needs to check if the move is in or out, which is not optimal since each valuation layer can have the same move.

### Fix:

We can't change the way `_account_entry_move` works in stable but we can batch `_is_in` and `_is_out`
outside the method. This change will reduce the time taken by each `_account_entry_move` and remove
unnecessary calls.

### Benchmark (in 18):

| # of moves          |  Before  |  After  |
|---------------------|----------|---------|
| 8 (with 1000 each)  |    13:57 |    3:46 |
| 3 (with 1489 each)  |     8min |    2:53 |
| 1 (with 1489)       |     2:47 |     36s |

### Reference:
opw-4550348