@robodoo 2025-02-27
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/199591.png)](https://mergebot.odoo.com/odoo/odoo/pull/199591)